+++ src/shaders/rope.wgsl
@@ -3,12 +3,12 @@ struct Params {
   dim : u32,
   L : u32,
   n_heads: u32,
-  base_freq: f16,
+  base_freq: f32,
   l_offset : u32,
   write_l_offset: u32,
 };
 
-@binding(0) @group(0) var<storage, read_write> K : array<f16>; //L * dim
+@binding(0) @group(0) var<storage, read_write> K : array<f32>; //L * dim
 @binding(1) @group(0) var<uniform> params : Params;
 
 @compute @workgroup_size(1)
@@ -29,8 +29,8 @@ fn main(@builtin(global_invocation_id) GlobalInvocationID : vec3<u32>) {
   var k : u32 = 0u;
   for (var h: u32 = 0u; h < params.n_heads; h = h + 1u) {
     for (var head_k : u32 = 0u; head_k < dim_per_head; head_k = head_k + 2u) {
-      let freq = 1.0 / pow(params.base_freq, f16(head_k) / f16(dim_per_head));
-      let val = f16(l) * freq;
+      let freq = 1.0 / pow(params.base_freq, f32(head_k) / f32(dim_per_head));
+      let val = f32(l) * freq;
       let real = cos(val);
       let img = sin(val);
